refactor: move database logic into repositories layer
