+++ backend/app/api/v1/endpoints/questions.py
@@ -4,11 +4,10 @@ from __future__ import annotations
 
 from fastapi import APIRouter, status
 
-from app.core.db_safe import execute_db_safe
 from app.core.deps import SupabaseSdkDep
 from app.models.questions import Question
+from app.repositories.questions import insert_question_row
 from app.schemas.questions import CreateQuestionRequest
-from app.utils.supabase_response import expect_single_row_from_result
 
 router = APIRouter(tags=["questions"])
 
@@ -21,9 +20,5 @@ router = APIRouter(tags=["questions"])
 )
 async def create_question(body: CreateQuestionRequest, client: SupabaseSdkDep) -> Question:
     payload = body.model_dump(mode="json", exclude_none=True, by_alias=True)
-    result = await execute_db_safe(client.table("questions").insert(payload).execute())
-    row = expect_single_row_from_result(
-        result,
-        detail="Unexpected response from Supabase when creating question.",
-    )
+    row = await insert_question_row(client, payload)
     return Question.model_validate(row)
