+++ backend/app/repositories/__init__.py
@@ -1,0 +1,1 @@
+"""Supabase PostgREST persistence (queries and writes)."""
